benchmark/database/requirements.txt
mysql-connector-python
psycopg2-binary
matplotlib
